Update the mechanism to check for new token.
It's now confirmed that new token is located in modified_config.

But problem is that the old token is still in tmp: remote, so the last version of the code overwrites original_config twice.

First overwrites it with new token, then reads old token from tmp: remote and overwrites it back to old token.

To solve the problem, the program now checks expiry time on tokens, only rewrites if new token is detected.